requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1